refactor(viz): rename expression viz module hail.py → anndata.py
The file was rewritten for AnnData during scverse migration but kept the
Hail-era filename. The leftover Hail histogram function had no live callers
(only @pytest.mark.hail tests against removed Hail builders) and a broken
``"ad.AnnData"`` annotation that triggered an F821 lint error.

- New hvantk/visualization/expression/anndata.py exposes
  visualize_expression_distribution(adata, ...) operating on adata.X.
- Old hail.py removed; visualization facade and subpackage doc updated.
- Test import switched to the new module.
