+++ hvantk/visualization/expression/__init__.py
@@ -1,11 +1,8 @@
 """
 Expression visualization namespace.
 
-This subpackage groups plots for expression-related data across backends and data sources.
-Current implementations:
-- hvantk.visualization.expression.hail: Hail MatrixTable support
-
-Additional backends (pandas/AnnData) may be added in future modules.
+Submodules:
+- ``hvantk.visualization.expression.anndata``: plots that operate on AnnData objects.
 """
 
-# Intentionally avoid importing heavy backends here; import from submodules directly
+# Intentionally avoid importing submodules eagerly to keep import lightweight.
